improve the interface
