+++ src/ma5_expert/__init__.py
@@ -12,7 +12,7 @@ log.setLevel(logging.INFO)
 from ma5_expert import cutflow
 from ma5_expert import histogram
 from ma5_expert import pad
-# from ma5_expert import backend
+from ma5_expert.backend import BackendManager, PADType
 
 
-__all__ = cutflow.__all__ + histogram.__all__ + pad.__all__ #+ backend.__all__
+__all__ = cutflow.__all__ + histogram.__all__ + pad.__all__ + ["BackendManager", "PADType"]
